moved printable to its own file
