+++ Include/Printable.h
@@ -1,0 +1,10 @@
+#include <any>
+#include <string>
+
+struct Printable {
+  /// @brief Wrapper around std::any to print variables of any type
+  Printable();
+  Printable(std::any value, std::string str);
+  std::any value;
+  std::string str;
+};
